requirements.txt
flask
flask-cors
python-dotenv
requests
gunicorn
Flask-SocketIO
paho-mqtt
